rename debug plugin name to logging plugin and add sql script.
