--- soul-spring-boot-starter/soul-spring-boot-starter-plugin/soul-spring-boot-starter-plugin-debug/src/main/java/org/dromara/soul/springboot/starter/plugin/debug/DebugPluginConfiguration.java
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.dromara.soul.springboot.starter.plugin.debug;
-
-import org.dromara.soul.plugin.api.SoulPlugin;
-import org.dromara.soul.plugin.debug.DebugPlugin;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * The type Debug plugin configuration.
- *
- * @author xuxd
- **/
-@Configuration
-public class DebugPluginConfiguration {
-
-    /**
-     * Debug plugin soul plugin.
-     *
-     * @return the soul plugin
-     */
-    @Bean
-    public SoulPlugin debugPlugin() {
-        return new DebugPlugin();
-    }
-
-}
