+++ soul-spring-boot-starter/soul-spring-boot-starter-plugin/soul-spring-boot-starter-plugin-logging/src/test/java/org/dromara/soul/springboot/starter/plugin/logging/LoggingPluginConfigurationTest.java
@@ -1,0 +1,51 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.dromara.soul.springboot.starter.plugin.logging;
+
+import org.dromara.soul.common.enums.PluginEnum;
+import org.dromara.soul.plugin.api.SoulPlugin;
+import org.dromara.soul.plugin.logging.LoggingPlugin;
+import org.junit.Test;
+import org.springframework.boot.autoconfigure.AutoConfigurations;
+import org.springframework.boot.test.context.runner.ApplicationContextRunner;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+/**
+ * Test case for {@link LoggingPluginConfiguration}..
+ *
+ * @author xuxd
+ **/
+public class LoggingPluginConfigurationTest {
+
+    @Test
+    public void testDebugPlugin() {
+        new ApplicationContextRunner()
+            .withConfiguration(
+                AutoConfigurations.of(
+                    LoggingPlugin.class
+                ))
+            .withPropertyValues("debug=true")
+            .run(
+                context -> {
+                    SoulPlugin plugin = context.getBean("loggingPlugin", SoulPlugin.class);
+                    assertThat(plugin.named()).isEqualTo(PluginEnum.Logging.getName());
+                }
+            );
+    }
+}
